src/healthcare_doc_processor/requirements.txt
# Core libraries
crewai>=0.19.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0

# Document processing
pytesseract>=0.3.10
pdf2image>=1.16.3
python-docx>=0.8.11
Pillow>=9.5.0

# NLP and embeddings
transformers>=4.35.0
sentence-transformers>=2.2.2
torch>=2.0.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# AWS services (for Textract)
boto3>=1.28.0

# Utilities
numpy>=1.24.0
pydantic>=2.0.0